Move isSpanFilter to filterValueFunctions.
